chore(refactor): create separate controllers for session requests and auth
