+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "start": "nodemon --exec index.js",
-    "test": "NODE_ENV= mocha  mocha tests/*.js --exit",
+    "test": "NODE_ENV=  mocha tests/*.js --exit",
     "coveralls": "NODE_ENV=test npm test && report --reporter=text-lcov | coveralls "
   },
   "repository": {
